[refactor][API] v2.0.0 FrontEND 구조화 변경 쿠폰 (admin/public) UI 와 스타일 통일화로 변경, 프로젝트 구조화 변경 Layer 단위로 폴더 구조 수정 REF : #11
