+++ web/src/domains/coupons/feature/admin/components/AdminCouponTable.jsx
@@ -1,12 +1,12 @@
 import React from "react";
-import CouponCreateModal from "@/domains/coupons/feature/components/admin/modal/CouponCreateModal.jsx";
+import CouponCreateModal from "@/domains/coupons/feature/admin/components/modal/CouponCreateModal.jsx";
 import AdminTableLayout from "@/global/layout/adminPageLayout/table/AdminTableLayout.jsx";
-import CouponTableHead from "@/domains/coupons/feature/list/admin/components/table/CouponTableHead.jsx";
-import CouponTableBody from "@/domains/coupons/feature/list/admin/components/table/CouponTableBody.jsx";
+import CouponTableHead from "@/domains/coupons/feature/admin/components/table/CouponTableHead.jsx";
+import CouponTableBody from "@/domains/coupons/feature/admin/components/table/CouponTableBody.jsx";
 import AdminFilterBar from "@/global/layout/adminPageLayout/table/AdminFilterBar.jsx";
-import CouponEditModal from "@/domains/coupons/feature/components/admin/modal/CouponEditModal.jsx";
-import { useAdminCouponTable } from "@/domains/coupons/feature/list/admin/hooks/useAdminCouponTable.js";
-import { useAdminCouponFilter } from "@/domains/coupons/feature/list/admin/hooks/useAdminCouponFilter.js";
+import CouponEditModal from "@/domains/coupons/feature/admin/components/modal/CouponEditModal.jsx";
+import { useAdminCouponTable } from "@/domains/coupons/feature/admin/hooks/useAdminCouponTable.js";
+import { useAdminCouponFilter } from "@/domains/coupons/feature/admin/hooks/useAdminCouponFilter.js";
 
 const AdminCouponTable = () => {
   const { coupons, changeVisible } = useAdminCouponTable();
