+++ web/src/domains/coupons/feature/admin/components/table/CouponTable.module.scss
