+++ web/src/domains/coupons/feature/admin/components/table/CouponTableBody.jsx
