+++ web/src/domains/coupons/feature/admin/hooks/useAdminCouponTable.js
@@ -1,17 +1,17 @@
 import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from "react";
-import { requestGetCouponList, requestUpdateCouponVisible } from "@/domains/coupons/store/index.js";
 import { VisibleToggleHandler } from "@/global/handler/VisibleToggleHandler.js";
+import { requestAdminUpdateCouponVisible, requestGetAdminCouponList } from "@/domains/coupons/store/admin/thunks.js";
 
 export const useAdminCouponTable = () => {
   const dispatch = useDispatch();
   const { coupons, loading, error } = useSelector((state) => state.coupon);
 
   useEffect(() => {
-    dispatch(requestGetCouponList());
+    dispatch(requestGetAdminCouponList());
   }, [dispatch]);
 
-  const changeVisible = VisibleToggleHandler(dispatch, requestUpdateCouponVisible);
+  const changeVisible = VisibleToggleHandler(dispatch, requestAdminUpdateCouponVisible);
 
   return {
     coupons,
