+++ web/src/domains/coupons/feature/admin/pages/AdminCouponListPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AdminCouponTable from "@/domains/coupons/feature/list/admin/components/AdminCouponTable.jsx";
+import AdminCouponTable from "@/domains/coupons/feature/admin/components/AdminCouponTable.jsx";
 
 const AdminCouponListPage = () => {
   return <AdminCouponTable />
