+++ web/src/domains/coupons/feature/public/components/CouponListSection.module.scss
