+++ web/src/domains/coupons/feature/public/components/card/CouponCard.module.scss
