+++ web/src/domains/coupons/feature/public/components/slide/CouponSlideSection.module.scss
