+++ web/src/domains/coupons/feature/public/utils/couponUtils.js
